api/requirements.txt
PrettyTable
requests
bottle
user-agents
unipath
google-cloud-pubsub
redis